Fix issue where admins couldn't see some challenges in the add requirements intereface
