+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# UNRELEASED
+
+**General**
+
+- Fix an issue where admins couldn't see challenges which had requirements
+
 # 3.3.0 / 2020-03-26
 
 **General**
